+++ link_analysis/api_module.py
@@ -1,18 +1,9 @@
-# coding=utf-8
-# Coded by Aleksandr Rodionov
-# rexarrior@yandex.ru
-
-
-# other imports---------------------------------------------------------
-import os.path
+import os
 import datetime
 import time
+
 import dateutil.parser
 # License: Apache Software License, BSD License (Dual License)
-
-# imports Core modules--------------------------------------------------
-from web_crawler import ksrf
-import web_crawler
 
 if __package__:
     from link_analysis import models
@@ -25,9 +16,7 @@ else:
     import converters
     import wc_interface
     import link_handler
-# methods---------------------------------------------------------------
 
-# internal methods------------------------------------------------------
 DECISIONS_FOLDER_NAME = 'Decision files'
 JSON_HEADERS_FILENAME = 'DecisionHeaders.json'
 JSON_HEADERS_FOR_CHECKING_LINKS_FILENAME = \
@@ -67,7 +56,6 @@ def load_and_visualize(pathTograph=PATH_TO_JSON_GRAPH):
 
 
 # api methods-----------------------------------------------------------
-
 
 def process_period(
         firstDateOfDocsForProcessing=None, lastDateOfDocsForProcessing=None,
@@ -410,78 +398,12 @@ def get_CODE_links_from_all_headers(
 
 if __name__ == "__main__":
     start_time = time.time()
-    # process_period("18.06.1980", "18.07.2020", showPicture=False,
-    #                sendRequestToUpdatingHeadersInBaseFromSite=False,
-    #                includeIsolatedNodes=True,
-    #                takeHeadersFromLocalStorage=True)
-    # process_period("18.06.1980", "18.07.2020", showPicture=False,
-    #                sendRequestToUpdatingHeadersInBaseFromSite=False,
-    #                includeIsolatedNodes=True,
-    # takeHeadersFromLocalStorage=True)
-    # process_period(
-    #     firstDateOfDocsForProcessing='18.03.2013',
-    #     lastDateOfDocsForProcessing='14.08.2018',
-    #     docTypesForProcessing={'КСРФ/О', 'КСРФ/П'},
-    #     firstDateForNodes='18.03.2014', lastDateForNodes='14.08.2017',
-    #     nodesIndegreeRange=(0, 25), nodesOutdegreeRange=(0, 25),
-    #     nodesTypes={'КСРФ/О', 'КСРФ/П'},
-    #     includeIsolatedNodes=False,
-    #     firstDateFrom='18.03.2016', lastDateFrom='14.08.2016',
-    #     docTypesFrom={'КСРФ/О', 'КСРФ/П'},
-    #     firstDateTo='18.03.2015', lastDateTo='14.08.2015',
-    #     docTypesTo={'КСРФ/О', 'КСРФ/П'},
-    #     weightsRange=(1, 5),
-    #     graphOutputFilePath=PATH_TO_JSON_GRAPH,
-    #     showPicture=True, sendRequestToUpdatingHeadersInBaseFromSite=False,
-    #     takeHeadersFromLocalStorage=True)
-
-    # start_process_with(decisionID='КСРФ/1-П/2015', depth=3)
-
-    # load_and_visualize()
-
-    # start_process_with(
-    #     decisionID='КСРФ/1-П/2015', depth=10,
-    #     firstDateForNodes='18.03.2014', lastDateForNodes='14.08.2018',
-    #     nodesIndegreeRange=(0, 25), nodesOutdegreeRange=(0, 25),
-    #     nodesTypes={'КСРФ/О', 'КСРФ/П'},
-    #     includeIsolatedNodes=False,
-    #     firstDateFrom='18.03.2011', lastDateFrom='14.08.2019',
-    #     docTypesFrom={'КСРФ/О', 'КСРФ/П'},
-    #     firstDateTo='18.03.2011', lastDateTo='14.08.2018',
-    #     docTypesTo={'КСРФ/О', 'КСРФ/П'},
-    #     weightsRange=(1, 5),
-    #     graphOutputFilePath=PATH_TO_JSON_GRAPH,
-    #     showPicture=True, sendRequestToUpdatingHeadersInBaseFromSite=False,
-    #     takeHeadersFromLocalStorage=True)
-
-    # source = web_crawler.Crawler.get_data_source('LocalFileStorage')
-    # text=source.get_data('КСРФ/19-П/2014',
-    #                       web_crawler.DataType.DOCUMENT_TEXT)
-    # text = wc_interface.get_text('КСРФ/1010-О-О/2008')
-    # process_period("01.09.2018", "18.07.2019", showPicture=True,
-    #                sendRequestToUpdatingHeadersInBaseFromSite=False,
-    #                includeIsolatedNodes=True,
-    #                takeHeadersFromLocalStorage=False)
-    # process_period("12.04.2018", "12.04.2018", showPicture=True,
-    #                sendRequestToUpdatingHeadersInBaseFromSite=False,
-    #                includeIsolatedNodes=True,
-    #                takeHeadersFromLocalStorage=False)
-    # cl1 = converters.load_pickle("Results0\сleanLinks.pickle")
-    # cl2 = converters.load_pickle("Results\сleanLinks.pickle")
-    # import my_funs
-    # my_funs.compare_clealinks(cl1, cl2)
-    # print(f"Headers collection spent {time.time()-start_time} seconds.")
-
-    # process_period(
-    #     nodesIndegreeRange=(0, 25), nodesOutdegreeRange=(0, 25),
-    #     graphOutputFilePath=PATH_TO_JSON_GRAPH,
-    #     showPicture=False, sendRequestToUpdatingHeadersInBaseFromSite=False,
-    #     takeHeadersFromLocalStorage=True)
 
     # response = get_all_links_from_all_headers()
     # converters.save_json(response, 'cleanLinks.json')
     # print(f"\nFound links: {len(response)}. "
     #       f"Time: {time.time()-start_time} seconds.")
+
     response = get_CODE_links_from_all_headers()
     print(f"\nFound links: {len(response)}. "
           f"Time: {time.time()-start_time} seconds.")
